# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Bank API
  version: 1.0.0
  description: API for managing bank accounts and transfers

servers:
  - url: http://localhost:3002/simplebank
    description: Development server

paths:
  /accounts:
    post:
      summary: Create a new account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '200':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: List all accounts
      operationId: listAccounts
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /accounts/{id}:
    get:
      summary: Get account by ID
      operationId: getAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Account ID
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete account by ID
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Account ID
      responses:
        '200':
          description: Account deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /accounts/transfers:
    post:
      summary: Create a new transfer
      operationId: createTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
      responses:
        '200':
          description: Transfer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    Account:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
          description: Unique account identifier
        owner:
          type: string
          description: Account owner name
        balance:
          type: integer
          format: int64
          description: Account balance in cents
        currency:
          type: string
          description: Currency code (e.g., USD)
      required:
        - accountId
        - owner
        - balance
        - currency

    CreateAccountRequest:
      type: object
      properties:
        owner:
          type: string
          description: Account owner name
        balance:
          type: integer
          format: int64
          description: Initial account balance in cents
        currency:
          type: string
          description: Currency code (e.g., USD)
      required:
        - owner
        - balance
        - currency

    Transfer:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique transfer identifier
        fromAccountId:
          type: integer
          format: int64
          description: Source account ID
        toAccountId:
          type: integer
          format: int64
          description: Destination account ID
        amount:
          type: integer
          format: int64
          description: Transfer amount in cents
      required:
        - id
        - fromAccountId
        - toAccountId
        - amount

    CreateTransferRequest:
      type: object
      properties:
        fromAccountId:
          type: integer
          format: int64
          description: Source account ID
        toAccountId:
          type: integer
          format: int64
          description: Destination account ID
        amount:
          type: integer
          format: int64
          description: Transfer amount in cents
      required:
        - fromAccountId
        - toAccountId
        - amount

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error

  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'